Fix IE bug when null passed to insertBefore
Fix a error occurring in IE when an undefined value was passed to
insertBefore.

`insertBefore(newElement, referenceElement)`: If referenceElement is
null, newElement is inserted at the end of the list of child nodes.
In Internet Explorer, an undefined value as referenceElement will
throw an "Invalid argument" exception, while in rest of the modern
browsers, this works fine.